Added a Step to Change Key permissions before SSH
Common error "It's required that your private keys are not accessible by others" and also, "Permission Denied (Public Key)" when trying to ssh the VM. As stated in the Azure portal, we should check the permissions through chmod first. Added instructions and code to do so before the ssh step.