Update storage-use-azcopy-configure.md
The older version of AzCopy 7.3 which is specifically for table storage has a different path than the newer version. Also, the file gets enabled only when Verbose switch is used else not. Hence proposing these changes to have the customer check the logs at the appropriate location in case they are using older version.