Different behavior than what is documented
In lines 47/48, the doc says "All other letters are converted to lowercase." But other letters even when typed in upper case is not converted to lower case. Either the doc is wrong or there is a issue in the behavior. The screen capture also does not show changes to lower case