Removed FAQ: "Why does the available space on my SMB client not show the provisioned size?"
Removed SMB FAQ entry because of hard quota enforcement:

### Why does the available space on my SMB client not show the provisioned size?

The volume size reported by the SMB client is the maximum size the Azure NetApp Files volume can grow to. The size of the Azure NetApp Files volume as shown on the SMB client is not reflective of the quota or size of the volume. You can get the Azure NetApp Files volume size or quota through the Azure portal or the API.