Update common-questions-appliance.md
Can I harden the appliance?
When you use the downloaded template to create the appliance, you can add components (antivirus, for example) to the template if you leave in place the communication and firewall rules that are required for the Azure Migrate appliance.

The part after "if" does not seem to make sense or the sentence is incomplete. I hope we can explain it in a better way as I could not really understand what it meant.