Update permissions-reference.md
Per testing and discussion Authentication Administrator role can't be utilized to reset MFA on owners or members of role assignable AAD security group.  This is mentioned in https://docs.microsoft.com/en-us/azure/active-directory/roles/groups-concept#why-we-enforce-creation-of-a-special-group-for-assigning-it-to-a-role but is not mentioned here and I believe it should be to prevent future confusion when troubleshooting this role's capabilities.