(AzureCXP) MicrosoftDocs/azure-docs#76571
In the actualRowCounts SUB-SELECT, the first filter in the WHERE clause is 1=1.
This should be WHERE rg.index_id = 1

Otherwise this will lead to duplicates when a table has more than one index. The results of the entire table will be off by a multiple of the number of indexes greater than one.