Provided resource group does not exist exception
Have seen quite a few support cases where resource group exists but Join-AzStorageAccountForAuth fails with the below confusing error. Its very difficult for support engineers to isolate the problem here with this error. This happens because customers grant access on storage account resource level only as storage account owner or contributor as mentioned in the article. But when Join-AzStorageAccountForAuth command is executed it validates that user's permission on resource group hosting this storage account also and since that user is not a reader it fails as resource group does not exist.

Get-AzResourceGroup : Provided resource group does not exist.

If the changes mentioned is posted customer will grant the required RBAC role referring the pre-requisites. Even if its missed, the error would be searched and the permission would be added.