Added note about asterisks in custom host headers
When attempting to use an asterisk in a ATM custom header for an endpoint/profile, ATM will not use that header and fall back to using the FQDN of the endpoint. This calls out that using an asterisk (sometimes incorrectly used when trying to access a multisite listener, for example) is unsupported and will not work.