Update logic-apps-securing-a-logic-app.md
Article https://docs.microsoft.com/en-us/azure/logic-apps/logic-apps-securing-a-logic-app?tabs=azure-portal#access-for-inbound-calls-to-request-based-triggers
Only lists as supported the cipher we actually recommend customers to use. But in fact the true list of supported ciphers is much longer. While we do not want customers to use the older ciphers, we had to keep them based on sustained usage to avoid breaking existing customers. i.e. backward compatibility. Although we do not list the older ciphers on the doc page, customers discover them by running security tools on our endpoints, and then open support tickets and security reports because they think our endpoints are unsecured.

We need to publish the full list of ciphers we support, while clearly stating that the older ciphers are not advised for any new use. They may even be retired in the future once the last customer has stopped depending on it.