Create durable-functions-singletons.md
Couple of things wrong with mainly this line:
```
if existing_instance != None or existing_instance.runtime_status in ["Completed", "Failed", "Terminated"]:
```

First: the runtime_status property is not a string, it's one of the fields in the OrchestrationRuntimeStatus Enum. Second, the first condition here says "if the existing instance does *not not* exist", which should be "if the existing instance does not exists".