SMTP is not supported but it still works well.
This document said "Outbound traffic on TCP port 25 isn't allowed". When I allowed outbound traffic on TCP port 25 in Azure Firewall, I connected to the Internet via Azure Firewall so the document was not correct. I know Azure Firewall does not support "Outbound traffic on TCP port 25" but the behavior is very confusing. Could you fix this document in order to avoid confusing ?