Update functions-event-grid-blob-trigger.md
It seems like the documentation on blob triggers and event grid triggers got smooshed together.
I had to change my invocation url in the local context to replace blob with event grid for it to work. 
For a python function, this url worked for me: http://localhost:7071/runtime/webhooks/eventgrid?functionName=FuncNameWithNoHostDotFunctionsInIt
I'm using the v3.0.3388 of the func tools.