Incorrect file size
The actual information displayed by Microsoft within the branding section is misleading/incorrect.  It says "File size: <300KB" which is 307,200 Bytes. When you have an image file that Windows shows is 299KB (306,176 Bytes) the Azure portal rejects the image as too large with an error displayed that says "Selected file <name> must be between 1 and 300000 bytes" which is less than 293KB.  I know this is not the correct way to get the actual portal functionality updated so let's update the documentation to reflect what the portal actually accepts.