Specify the US locale for Java sample and clarify date format
If the locale on your machine is not US, then the date gets generated differently and Azure Monitor Data Collector API rejects the message with an invalid date format message.

I'm in Australia and the date format ended up looking like `Fri., 02 Jul. 2021 07:28:21 GMT` which has extra full stops after the day and month. Adding the locale fixed that so the date became `Fri, 02 Jul 2021 07:29:15 GMT`.

Ultimately I switched to `DateTimeFormatter` because it's more modern and performant (it's thread safe so you can share a formatter between threads). But it's a little fiddly to get to work. 
Out of the box there is a formatter that almost works, `DateTimeFormatter.RFC_1123_DATE_TIME`, but it doesn't have leading zeros for the day of the month (I'm fairly certain RFC 1123 allows for 1 or 2 digits for day of month), which causes the API to fail. Since the Data Collector API is quite strict about the dates, I think it should say that the date format is [RFC 7231](https://datatracker.ietf.org/doc/html/rfc7231#section-7.1.1.1) instead of [RFC 1123](https://datatracker.ietf.org/doc/html/rfc1123#page-55). RFC 7231 is the same format used for HTTP Expires headers, which is more in line with how the header is used.