Update refresh-tokens.md -> Token timeouts
Doc states "Using token lifetime configuration, the lifetime of refresh tokens can be reduced or lengthened. ", but this no longer verifies, as it is not possible anymore to configure token lifetimes for refresh tokens (https://docs.microsoft.com/en-us/azure/active-directory/develop/active-directory-configurable-token-lifetimes#token-lifetime-policies-for-refresh-tokens-and-session-tokens).
Due to this, information is outdated, and might be replaced with the info in the aforementioned link, pointing to the new way of doing it and to information about default lifetime properties for refresh tokens.