Clarifying this in more detail
Phone call verification was not clearly identified being in that paragraph.  By placing it in a note, and providing an example it makes it clearer.  Most of our customers will know they have a trial license not a trial subscription even though the backend shows it as a subscription. By providing the example, this helps clarify that a trial license will not work with phone call verification either.