Updated Recovery Point Objective (RPO) details
Subsection Service-level objectives is updated with the following -- 
Recovery Point Objective (RPO) indicates the point in time to which data can be recovered. The RPO target is typically less than twice the replication schedule, but it can vary. In some cases, it can go beyond the target RPO based on factors such as the total dataset size, the change rate, the percentage of data overwrites, and the replication bandwidth available for transfer.


•	For the replication schedule of 10 minutes, the typical RPO is less than 20 minutes.
•	For the hourly replication schedule, the typical RPO is less than two hours. 
•	For the daily replication schedule, the typical RPO is less than two days.