Global Catalog access is needed to enable Azure Active Directory Seamless Single Sign-On
Without access to the GC the scripts to enable Azure Active Directory Seamless Single Sign-On seen at https://docs.microsoft.com/en-us/azure/active-directory/hybrid/tshoot-connect-sso#manual-reset-of-the-feature do not work because the cmdlets report "[time] [  9] [INFORMATIONAL] GetDesktopSsoComputerAccount: Searching in global catalog(forest) and ad.dcsadvisory.com for computer account AZUREADSSOACC" and then fail to connect. Once GC ports are open this works immediately. Therefore proposing that this is added to this table