Fix out-of-date docs for AAD /.auth/refresh
The current AAD instructions for enabling refresh scenarios assumes the
user is configured with the classic Authentication configuration
expreince and that they are using an AAD v1 issuer, both of which are no
longer the default experience.

The new instructions work with both AAD v1 and v2 issuers, and the
primary instructions are for authsettingsV2, with a NOTE for customers
still using authsettings.