Chschrae/intent recognizer pattern updates (#68)
* changed the initial intent recognition page

* Made some grammatical changes (#67)

To remove passive voice and be a little more clear. Nothing major.
One thing is missing:
When would I use the built in intents on the Speech side, and when would I use LUIS. You say it to a degree, but the line between the two isn't clear. It would be great if we could provide more guidance here.

* typo fix

* Added recommendation to start with pattern matcher

* minor change

Co-authored-by: oliver scholz <oliversc@microsoft.com>