Update forced-tunneling.md
Reworded for clarity because at the moment one paragraph states to have route propogation disabled on AzureFirewallSubnet and AzureFirewallManagementSubnet , the next paragraph states to enable route propogation on AzureFirewallSubnet and this is confusing. 
Also changed the sentence "BGP route propogation" to current format "propogate gateway routes" to make it consistent.