Update Timer description
The Timer trigger does not execute at a 'set time', it executes with a given CRON expression or timespan. Therefore the result is a given period (or interval) between each execution. Semantics maybe - but I have had to explain this to a few end users, that this is not a scheduling agent trigger that will run a task at a specific time.