Added FAQ "Which Unicode Character Encoding is supported by Azure NetApp Files for the creation and display of file and directory names?"
Added under Product FAQs:

### Which Unicode Character Encoding is supported by Azure NetApp Files for the creation and display of file and directory names?   

Azure NetApp Files only supports file and directory names that are encoded with the UTF-8 Unicode Character Encoding format for both NFS and SMB volumes.

If you try to create files or directories with names that use supplementary characters or surrogate pairs such as non-regular characters and emoji that are not supported by UTF-8, the operation will fail. In this case, an error from a Windows client might read “The file name you specified is not valid or too long. Specify a different file name.”