Update custom classification minimum match threshold
There are no false-positives when it comes to custom classification.
We ask the customer to define the percentage match they need for the custom classification to be applied.
If match % >= customer defined %, the tag is applied.
If match % < customer defined %, the tag is not applied.

So, there are no false positives here. We should let the customers define whatever match % they want.