Instructions match current UI (Windows instead of Azure)
We instruct readers to search for "Azure Virtual Desktop" after Windows Virtual Desktop was renamed, however this is still "Windows Virtual Desktop" in the portal.  In order to complete the tutorial, readers need to search for "Windows Virtual Desktop" ("Azure Virtual Desktop" returns no results).

This will need to be updated as soon as the portal is updated.  I've added a bit of text to minimize reader impact during the delay between the portal update and the doc update ("this will soon be updated to "Azure Virtual Desktop").

This resolves https://github.com/MicrosoftDocs/azure-docs/issues/79415.