Use of WEBSITE_DISABLE_SCM_SEPARATION with Local Cache
Noticed App Service showing deployed content immediately on the App Service without a restart when using Local Cache, in the presence of the App Setting - 'WEBSITE_DISABLE_SCM_SEPARATION = true'. If the App restarts in such a case - it reverts to files before the deployment, since the deployed files are not written to persistent storage.