(AzureCXP) fixes MicrosoftDocs/azure-docs/issues/79515
Updated this Statement to : Contains all intermediate certificate and the root one in the certificate chain.

Reference: https://docs.microsoft.com/en-us/azure/app-service/configure-ssl-certificate#merge-intermediate-certificates

" Create a file for the merged certificate, called mergedcertificate.crt. In a text editor, copy the content of each certificate into this file. The order of your certificates should follow the order in the certificate chain, beginning with your certificate and ending with the root certificate."