Expanded HSM acronym in What is section
Nowhere in the article do we explain what the HSM acronym stands for. I added Hardware Security Modules in brackets to the What is section, but feel free to alter if you have better formatting. 

The Dedicated HSM has a nice short HSM explainer in the FAQs, which could also be linked to or re-used here? https://docs.microsoft.com/en-us/azure/dedicated-hsm/faq

-------
cc: @msmbaldwin