Update manage-create-arc-environment.md
Remove LogAnalyticsKeyWithSpace line of PowerShell as handling whitespace is not required in PowerShell and current instructions was resulting in an empty string value in $LogAnalyticsKeyEnc as a result and this caused the extension deployment to fail due to not being able to authenticate with Log Analytics successfully.  Whitespace handling is only required for Bash.