Moved --enable-addons monitoring parameter to correct section
Step 2 of "Enable cluster monitoring" is not actually a step that can be taken as it is part of the command in the following section.  This moves the description of enabling Azure Monitor for containers into the section where the action is taken making it more clear both what needs to be done by the reader and what they should expect if they click on the link.  This resolves https://github.com/MicrosoftDocs/azure-docs/issues/79615.