Clarify note about the billing impact of queue polling
I encountered a case where the note about how polling is "free" was misinterpreted to mean that there is no charge to the Azure Storage account. I think the problem is that the text as written does not making it clear what "account" is being referred to. I decided to rewrite it to be a bit more explicit, and put it in a `[!NOTE]` since it only applies to the Consumption plan.