Added known limitations and post host pool deploy
Consolidated some of the known limitations into a section and tried to highlight/ consolidate post Host Pool configuration for access that is in addition to the usual.  

Removed 2 note boxes under "Access Azure AD-joined VMs - added to newly created "Known Limitations" section
Added "Known Limitations" section and consolidated any of those noted throughout the document
Post Host Pool deployment added a 3rd bullet regarding the non-Azure AD-joined clients and need for additional RDP Property. 

As a FastTrack Engineer these highlights will make it easier to quickly answer the questions on whether this solution fits the customer scenario.  Some of the consolidated bullets are noted throughout the document but not easily visible in a centralized section.