Update backup-rbac-rs-vault.md
Under the heading minimum role requirements for the Azure File share backup, for the operation to enable backup of Azure file shares, the document lists "Storage Account" in the role required column and "Contributor Storage account resource" in the resources column.

The role required should be "Storage Account Contributor" and the resource should be "Storage account resource". This change moves the third pipe on that line to fix it.