Update resize-vm.md
Just gave a try to Get-AzVmSize - and I'm getting a huge list of VM Sizes. Given it's not possible that a given cluster supports all of them (from A-series, to N-series, to M-series), the Get-AzVMSize returns the list of VM sizes available in the region - not in the cluster.
Doc needs to be udpated.