command update on pulling runtime container
Under the section "Store runtime containers in your private registry", the docker pull command "docker pull mcr.microsoft.com/azureiotedge-agent" is not accurate now. This command will auto refer to the version "latest", and it will fails to find the container as the "latest" version is not exist. suggested to change to "docker pull mcr.microsoft.com/azureiotedge-agent:[version]". and suggest user to find the latest version from azureiotedge-agent description page. 
This container has a latest version 1.0 which must be specified during pull, as I have checked that there is a policy changes of microsoft container version (please correct me if this is not the latest policy), that there will be no default "latest" version, and must specify the version number for every release.