Update - Include Publish Branch Description
The configuration pane in Azure Data Factory for Azure DevOps shows an entry for the Publish Branch. The table that describes each setting is missing an entry for the Publish Branch and what it is. It would be helpful to include a row in the table to indiciate to users what the Publish Branch is and what the default entry signifies.