JIT requires classic rules with FW
While JIT will not work with Firewalls managed with Firewall Manager, the Firewall must also be managed via Rules (classic) as there is an option to still deploy a Firewall with Policies and not explicitly use the Firewall Manager.  The NOTE as is implies that either mode of the firewall (Policy or classic) would work, but he Policy version also fails.