Update build-first-body-app.md
Use "sensor_capture" object name instead of "capture" in the section "Get captures from the Azure Kinect device", since sensor_capture is the object name used in the following section (and in the full code at the end of the page). Also fix missing declaration of sensor_capture (when I read it first, I did not know where this object came from).