Update ilogger.md
With the asp.net core console application, the current code only addresses how to use ILogger in Program. cs and the proposed code discuss how to use ILogger in the different classes via dependency injection which was missing. This proposed code will allow the other classes to use ILogger and log to the Application Insights.