Config/code changes required for feature flag
There are two minor changes that are required to get the example to run successfully including a prefix change to Beta in HelloController.java and an additional configuration line in bootstrap.properties to enable the feature flags as they are disabled by default.  This resolves https://github.com/MicrosoftDocs/azure-docs/issues/81125.