join-ubuntu-linux-vm.md
The ntpd is deprecated and the most recent version of Ubuntu support chronyd. The ntp configuration is outdated.
Moreover, trusting the reliability of the Azure platform, now the VM are synched with the host, it's possible to consider all the servers into the domain synched even if not directly synched with the domain controllers. For this reason, no additional configuration is required for the Network Time Protocol.