Update iot-support-help.md
Add the following note:
If you are asking about an issue on a new support request that happened more than a certain period ago, it is unlikely that we will still have logs to investigate. Please let us know more quickly in the future if you observe the issue again.