Fixed numbering of steps for creating peering with Azure portal
I noticed the step number referred to in later steps were wrong. When I edited the document to fix that it turned out that we were skipping quite a few numbers in the code (which are then corrected in the final document), so although the number referred to in the source code was correct, the resulting final document was wrong.

Both step numbers and the numbers referred to in the text have been fixed.