Fix in script "add UDR rules to Route Table"
This commit fixes a problem that the IP addresses of service tag "BatchNodeManagement" are not found. The output of `Get-AzNetworkServiceTag` does not include "BatchNodeManagement.WestEurope" (or any other location). By removing the location it returns the IP addresses that are necessary for the routing table.

The other changes are suggestions from the PSScriptAnalyzer.