Update concept-primary-refresh-token.md
Need to change the link in the Next Step to point to certain part of the article to eliminate confusion about how to run dsreccmd /status to get user PRT. Old link was pointing to the beginning of the article (where we instruct to run it as admin to troubleshoot device state), but we need to point to the section where we dsregcmd to troubleshoot AAD PRT for current user. 

Please correct the link formating if it was submitted incorrectly.