Updates to howto-migrate-single-flexible-minimum-downtime.md
Data-in replication along with mydumper/ myloader can also be used to do external migrations. We do not have separate doc explaining it in detail and this particular documentation is for single to flex server migration. 
So added an update to this doc which calls out that this approach can also be used to do external migrations.