update the Important note with python dependencie
I recently try to run this quickstart and my experience was very challenging and time consuming, all because this dependency to an old and unsupported version of python. I will like to suggest this change in order to avoid that others users like me go through this experience and enjoy our turn key service.