Update lifecycle-management-overview.md
There should be an extra '/' at end of the prefix match. If not then in case there is a blob name blob1abc inside the container sample-container1  then even that will get picked up by the policy as the prefix match is for sample-container/blob1.  Hence, proposing the changes for review

I tried proposing this under below PR as well however the example now seem to have changed (file name) but still not correct
https://github.com/MicrosoftDocs/azure-docs/pull/73290