Minor text changes in the VM section
For the table under step 2 in the Virtual Machine section, I made the following changes to reflect the choices presented in the portal when making a VM: 
Availability options: Select No infrastructure redundancy required. (previously said 'Leave the default of no redundancy required."
Image: Select Windows Server 2019 Datacenter - Gen2 (previously was set to select Gen1 which is not a selectable option).