Add link to Mapping Data Flows
Poor click through analysis revealed we have a lot of traffic to mapping data flows in the Next Steps at the very bottom of the page, and a lot of clicks in-page to the Mapping Data Flow properties section much higher up on the page, which has no direct link to the Mapping Data Flows article though.  Hopefully adding this link should improve click-through.  My theory is most people are looking for Mapping Data Flows article but many may click to that section, not find it there, and end up navigating away with search, etc.
-------
cc: @jianleishen