Adding clarity and examples
Customers tend to use the examples within this literally, and do not want to necessarily build policies using TAP, especially since it's in Preview. I resurrected the old examples that defined network location so that it shows it can be used for either employees or guests, and also built-out an example for device compliance using Filter for devices. Also added some clarity that organizations may need to build a series of policies relying on some or all of the examples depending on their user personas. #ATCP