Known issue with NTLMv2
We need to add the follow information within the FAQ document: 

What are the possible reasons why a user(s) is failing with the error code: LsaLogonUser failed with NTSTATUS -1073741715 for MFA Server. 

Error 1073741715 = Status Logon Failure -> The attempted logon is invalid. This is either due to a bad username or authentication. 

A plausible reason for this: If the primary credentials entered are correct, there could be a mismatch between the supported NTLM version on the MFA server vs. the Domain Controller. MFA Server only supports NTLMv1 (LmCompatabilityLevel=1 thru 4) and not NTLMv2 (LmCompatabilityLevel=5)