Need to add the warning below
The machine will not be accessible after creating image.  
If you generalize the virtual machine, no further actions on it are allowed. For example, a generalized virtual machine can't be started or modified. You can select this checkbox to have the virtual machine deleted automatically once the image is created. If you do not generalize the virtual machine, you can continue to use it as normal.