Update how-to-visual-studio-develop-module.md
I see a lot of new IoT Edge Module Developers get sidetracked or bottlenecked because they are not pointing to the correct Docker Mode (Linux containers vs Windows Containers mode). We should emphasize in the documentation/article that to save yourself some time and effort please make sure you are pointing to the correct Docker Container mode. I know we mention this in the Prerequisites section, but we should re-emphasize it again in the Building and Debugging sections later in the article.
Thanks, and let me know if you have any questions.
-Andres