Updated 'AccountName' to 'StorageAccountName' to avoid confusion.
In the last step of System-Assigned Identity association with Function. AccountName is used, which is a bit confusing, I changed it to 'StorageAccountName' because only Storage Account Name is required. Article need another update, because it is mentioned as following:

## Use managed identity for AzureWebJobsStorage (Preview)

Next you will use the system-assigned identity you configured in the previous steps for the `AzureWebJobsStorage` connection. 

But in this article, author showed to configure user-assigned identity not 'system-assigned identity'. The above line should have link of article, where 'system-assigned-identity' creation is shown.