Update management-groups-roles.md
Edited sentence references an administrator as being "it"...

"Once the Azure roles have been assigned to the users, the tenant administrator should remove _**itself**_ from the user access administrator role."

Themself would be a better choice of reflexive pronoun and is considered a proper non-gender specific term for humans.  If the preference is to indicate all living beings are together in this world and we are all things, no changes are required.