Capital of Ukraine is mispelled
The correct spelling of Ukrainian capital is not Kiev, it is Kyiv.   Ukrainian Ministry of Foreign Affairs (MFA) even launched a campaign called KyivNotKiev
Transliteration Kyiv was legally approved by the Ukrainian government in 1995. Kiev transliteration is a linguistic relic of Soviet occupation of Ukraine and should not be used any more. 
Thank you