Update for Username claim value
Update to clarify the recurring issue clients have been having with the username attribute claim. For new clients they would want to pass the UPN value but existing clients may be using the on-prem samaccountname or another attribute from AD. This username mismatch will cause duplicate accounts in the PolicyStat user database when the user authenticates and they will want to ensure the values match.