Adding hyperlink to updated article.
Adding hyperlink to updated article at the top of Next Steps section. Please consider updating the content on the two links following this new one. It can be a bit confusing for the reader. Maybe the second link (line 52) should just redirect to the one I am adding here (line 51).