Update upgrade-to-data-lake-storage-gen2-how-to.md
In case there is a feature which is not supported in ADLS Gen 2 or is in Preview  e.g. Soft Delete for Blobs as per current scenario, validation would fail. Making a pointer in the documentation and adding a link so that customers can verify based on exception too. Please review