(AzureCXP) MicrosoftDocs/azure-docs#83954
Changed line 45 to clarify the difference between stopped(deallocated) and stopped power states as below:

Dynamic addresses are assigned after a public IP address is associated to an Azure resource and is started for the first time. Dynamic addresses can change if a resource such as a virtual machine is stopped (deallocated) and then restarted through Azure. The address remains the same if a virtual machine is rebooted or stopped from within the guest OS. When a public IP address resource is removed from a resource, the dynamic address is released.

Fixes MicrosoftDocs/azure-docs#83954

-------
cc: @asudbring
GitHub Issue: https://github.com/MicrosoftDocs/azure-docs/issues/83954