Update multi-user-authorization.md
Added some clarification.
I think the issue arises from the fact that a new separate tenant would not be using the Recovery Services Vault, so the provider never gets registered. 
If the resource guard was on the same tenant as the Recovery Services Vault, then this would not be an issue, is it would have auto registered the provider during initial provisioning.