Update faq.yml
vSAN is always thin, even when a storage policy is set to object space reservation of 100%.  Data is not consumed until it is written.  However with OSR enabled, it does "reserve" the space on the capacity tab for data management purposes and will not allow you to provision more if you have OSR's that exceed your capacity.