Update note on Last RBAC Admin
Customer deleted the last RBAC owner using Tenant Admin with User Access Admin role on subscription and raised a concern that there is contradicting statement on document which states we do not allow deletion of last RBAC Admin as it is not supported. Later when communicated with PG was informed that this is possible only if tenant admin is making the delete call. Customer has requested to have this documented to avoid confusion and provide some more clarity.

Supporting PG WorkItem statement for proposed change: https://portal.microsofticm.com/imp/v3/incidents/details/274958788/home