Update service-accounts-group-managed.md
Fixed an incorrect assertion that the Managed Service Accounts container is an Organizational Unit (OU).  This is, in fact, not an OU but an Active Directory container; meaning that it is a different object type.  One key differentiator is that AD containers cannot have group policies linked to them.  They can only be linked to OUs.