Update hybrid-cloud-to-on-premises.md
I've updated the older sample script and it should good for production use. We used to have this script https://github.com/dstefanMSFT/active-direcotry-b2b-user-pull as a sample, but it's been 5 years and needed to be updated. The new script has the following improvements:
Using the Microsoft Graph PowerShell module instead of AzureADPreview
Updating App Registration to use MS graph instead of Azure AD graph permissions
Use certificate-based authentication instead of app secret
Added a "What if" functionality
Added capability for guests to be re-added to the Azure AD group and they will be restored (used to just fail)