Update high-cpu-diagnose-troubleshoot.md
Added the following under "Common causes of new and unusual high CPU utilization are"

* Query plan regression due to out of date statistics after modifications from insert, update, delete, or merge operations change the data distribution in the table or indexed view.

I see this to be a common pattern in the field. With the older compatibility mode (<130)for Azure SQL Database, the newer decreasing, dynamic statistics recompilation threshold that adjusts according to the table cardinality at the time statistics were evaluated is not applied and that poses a problem for larger tables.